LayoutTests/http/tests/websocket/tests/hybi/unpaired-surrogates-in-message-expected.txt

CONSOLE MESSAGE: Websocket message contains invalid character(s).
Checks whether SYNTAX_ERR is thrown when attemping to send unpaired surrogates.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS SYNTAX_ERR was thrown.
PASS successfullyParsed is true

TEST COMPLETE
